# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: School Library Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all students
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: graduationYear
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      summary: Add a new student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Created student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
        "404":
          description: Not found
    patch:
      operationId: updateStudent
      tags: [students]
      summary: Update a student
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete a student
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /students/{id}/loans:
    get:
      operationId: getStudentLoans
      tags: [students]
      summary: Get all loans for a student
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Student loans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Loan"

  # ─── Loans ────────────────────────────────────────────────────────────────
...
  /reservations:
    get:
      operationId: listReservations
      tags: [reservations]
      summary: List all reservations
      parameters:
        - name: studentId
          in: query
          schema: { type: integer }
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, fulfilled, cancelled]
      responses:
        "200":
          description: List of reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"
    post:
      operationId: createReservation
      tags: [reservations]
      summary: Reserve a book
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationInput"
      responses:
        "201":
          description: Created reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
  /reservations/{id}:
    patch:
      operationId: updateReservation
      tags: [reservations]
      summary: Update reservation status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationUpdate"
      responses:
        "200":
          description: Updated reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
    delete:
      operationId: deleteReservation
      tags: [reservations]
      summary: Cancel a reservation
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Cancelled

  # ─── Fines ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Student:
      type: object
      required: [id, name, email, studentId, graduationYear, borrowLimit, activeLoansCount, createdAt]
      properties:
        id:             { type: integer }
        name:           { type: string }
        email:          { type: string }
        studentId:      { type: string }
        phone:          { type: ["string", "null"] }
        graduationYear: { type: integer }
        borrowLimit:    { type: integer }
        activeLoansCount: { type: integer }
        totalFinesOwed: { type: number }
        createdAt:      { type: string }
    StudentInput:
      type: object
      required: [name, email, studentId, graduationYear]
      properties:
        name:           { type: string }
        email:          { type: string }
        studentId:      { type: string }
        phone:          { type: string }
        graduationYear: { type: integer }
        borrowLimit:    { type: integer }
    StudentUpdate:
      type: object
      properties:
        name:           { type: string }
        email:          { type: string }
        phone:          { type: string }
        graduationYear: { type: integer }
        borrowLimit:    { type: integer }
    Loan:
      type: object
      required: [id, studentId, bookId, checkedOutAt, dueDate, status]
      properties:
        id:           { type: integer }
        studentId:    { type: integer }
        bookId:       { type: integer }
        checkedOutAt: { type: string }
        dueDate:      { type: string }
        returnedAt:   { type: ["string", "null"] }
        status:       { type: string, enum: [active, returned, overdue] }
        studentName:  { type: ["string", "null"] }
        bookTitle:    { type: ["string", "null"] }
        bookAuthor:   { type: ["string", "null"] }
...
    Reservation:
      type: object
      required: [id, studentId, bookId, status, createdAt]
      properties:
        id:          { type: integer }
        studentId:   { type: integer }
        bookId:      { type: integer }
        status:      { type: string, enum: [pending, fulfilled, cancelled] }
        createdAt:   { type: string }
        studentName: { type: ["string", "null"] }
        bookTitle:   { type: ["string", "null"] }
        bookAuthor:  { type: ["string", "null"] }
    ReservationInput:
      type: object
      required: [studentId, bookId]
      properties:
        studentId: { type: integer }
        bookId:    { type: integer }
    ReservationUpdate:
      type: object
      properties:
        status: { type: string, enum: [pending, fulfilled, cancelled] }
...