# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * School Library Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  name: string;
  email: string;
  studentId: string;
  /** @nullable */
  phone?: string | null;
  graduationYear: number;
  borrowLimit: number;
  activeLoansCount: number;
  totalFinesOwed?: number;
  createdAt: string;
}