# lib/api-zod/src/generated/types/studentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * School Library Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentInput {
  name: string;
  email: string;
  studentId: string;
  phone?: string;
  graduationYear: number;
  borrowLimit?: number;
}